pyproject.toml
[project]
name = "finance-gpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "black>=25.1.0",
    "bpython>=0.25",
    "finnhub-python>=2.4.24",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.7",
    "langchain-mongodb>=0.6.2",
    "langdetect>=1.0.9",
    "openai>=1.95.0",
    "pydantic>=2.11.7",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.1",
    "reflex>=0.8.1",
    "rich>=14.0.0",
    "streamlit>=1.46.0",
    "tqdm>=4.67.1",
]
